anm productos y comienzo de clientes

+++ rentabike/src/app/abms/productos/listado-productos/listado-productos.component.ts
@@ -1,5 +1,7 @@
 import { ProductoService } from './../../../gestion/producto.service';
 import { Component, OnInit } from '@angular/core';
+import Swal from 'sweetalert2/dist/sweetalert2.js'
+
 
 @Component({
   selector: 'app-listado-productos',
@@ -10,18 +12,60 @@ export class ListadoProductosComponent implements OnInit {
 
   constructor(private _productosService:ProductoService) { }
   productos:any;
+  termino = "";
+  backup:any;
   ngOnInit(): void {
     this._productosService.getBikes().subscribe(data => {
       this.productos = data;
+      this.backup = this.productos;
     });
   }
 
   eliminarProducto(id){
-    this._productosService.deleteBike(id).subscribe(response=>{
-      let newProductos = this.productos.filter(item =>{
-        return item.id !== id
-      });
-      this.productos = newProductos;
+
+    Swal.fire({
+      title: 'Estas seguro?',
+      text: 'Estas a punto de eliminar un producto!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Si!',
+      cancelButtonText: 'No'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this._productosService.deleteBike(id).subscribe(response=>{
+          let newProductos = this.productos.filter(item =>{
+            return item.id !== id
+          });
+          this.productos = newProductos;
+          Swal.fire(
+            'Eliminado!',
+            'El producto ha sido eliminado.',
+            'success'
+          );
+        });
+
+      // For more information about handling dismissals please visit
+      // https://sweetalert2.github.io/#handling-dismissals
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        Swal.fire(
+          'Cancelado',
+          'El producto no sera eliminado :)',
+          'error'
+        )
+      }
     })
+  }
+
+  filtrar(){
+    let newProductos = this.productos.filter(producto => {
+      return producto.nombre.toLowerCase().includes(this.termino.toLowerCase()) === true
+    });
+    this.productos = newProductos;
+  }
+
+  handlerInput(event){
+    if(event.target.value.length === 0){
+      this.productos = this.backup;
+    }
   }
 }
